Relocated LoadLibrary into OspreySharp.IO
* Moved LoadLibrary from AbstractScoringTask to new OspreySharp.IO.LibraryLoader (its collaborators already live in IO); logging injected via Action<string> callbacks since it logged through _ctx (not a verbatim move)
* Removed the now-redundant pwiz.OspreySharp.IO using from AbstractScoringTask (LoadLibrary was its only IO consumer); single call site updated to LibraryLoader.Load
* Logic and log messages unchanged: Stellar 1e-9 cross-impl bit-parity PASS, 345/347 tests, inspection 0/0

See ai/todos/active/TODO-20260530_ospreysharp_loadlibrary_to_io.md
